Move errors and helpers into errors.go
